chore: rename //src/deno_internal.h to //src/internal.h
