+++ BUILD.gn
@@ -69,7 +69,7 @@ static_library("libdeno") {
 v8_source_set("deno_nosnapshot") {
   sources = [
     "src/binding.cc",
-    "src/deno_internal.h",
+    "src/internal.h",
     "src/file_util.cc",
     "src/file_util.h",
     "src/deno.h",
