+++ src/binding.cc
@@ -28,7 +28,7 @@ IN THE SOFTWARE.
 #include "third_party/v8/include/v8.h"
 #include "third_party/v8/src/base/logging.h"
 
-#include "./deno_internal.h"
+#include "internal.h"
 #include "deno.h"
 
 namespace deno {
