+++ src/from_snapshot.cc
@@ -8,7 +8,7 @@
 #include "third_party/v8/include/v8.h"
 #include "third_party/v8/src/base/logging.h"
 
-#include "./deno_internal.h"
+#include "internal.h"
 #include "deno.h"
 
 #ifdef DENO_MOCK_RUNTIME
